+++ tools/make-toc.js
@@ -2,7 +2,7 @@
 var generateDocs = require('./generate-docs');
 
 generateDocs({
-  src: './core.js',
+  src: './index.js',
   output: './markdown/generated-toc.md',
   template: './tools/template-toc.md.ejs'
 });